gpu issue

+++ test.py
@@ -62,7 +62,7 @@ def head_detect(_tracker,f_num,df,img0,_model,_device,
     pred = non_max_suppression(pred, conf_thres, iou_thres, None, False, max_det=max_det)
     for j, det in enumerate(pred): 
         det[:, :4] = scale_coords(img.shape[2:], det[:, :4], img0.shape).round()
-        cpu=det.numpy().float()
+        cpu=det.numpy().astype(float)
         tracks=_tracker.update(cpu,img0)
     xyxys = tracks[:, 0:4].astype('int') # float64 to int
     normals=xyxys / shape_mat
